*** empty log message ***
git-svn-id: https://svn.library.emory.edu/svn/slavevoyages/trunk@942 2440b244-9809-4ec7-bebd-234906d9abf5

+++ src/edu/emory/library/tast/dm/Slave.java
@@ -31,7 +31,7 @@ public class Slave {
 	
 	private long id;
 
-	private long voyageId;
+	private int voyageId;
 
 	private int age;
 
@@ -141,11 +141,11 @@ public class Slave {
 		this.source = source;
 	}
 
-	public long getVoyageId() {
+	public int getVoyageId() {
 		return voyageId;
 	}
 
-	public void setVoyageId(long voyageid) {
+	public void setVoyageId(int voyageid) {
 		this.voyageId = voyageid;
 	}
